Fix: Improve code structure in pair-programming with Lyam
